Move SnapshotProcessController into a separate file
