+++ AWSAppSyncClient/AWSAppSyncClient.swift
@@ -42,69 +42,6 @@ enum AWSAppSyncGraphQLOperation {
 
 internal let NoOpOperationString = "No-op"
 
-class SnapshotProcessController {
-    let endpointURL: URL
-    var reachability: Reachability?
-    private var networkStatusWatchers: [NetworkConnectionNotification] = []
-    let allowsCellularAccess: Bool
-
-    init(endpointURL: URL, allowsCellularAccess: Bool = true) {
-        self.endpointURL = endpointURL
-        self.allowsCellularAccess = allowsCellularAccess
-        reachability = Reachability(hostname: endpointURL.host!)
-        reachability?.allowsCellularConnection = allowsCellularAccess
-        NotificationCenter.default.addObserver(self, selector: #selector(checkForReachability(note:)), name: .reachabilityChanged, object: reachability)
-        do {
-            try reachability?.startNotifier()
-        } catch {
-        }
-
-        NotificationCenter.default.addObserver(self, selector: #selector(SnapshotProcessController.checkForReachability), name: NSNotification.Name(rawValue: kAWSDefaultNetworkReachabilityChangedNotification), object: nil)
-    }
-
-    @objc func checkForReachability(note: Notification) {
-
-        let reachability = note.object as! Reachability
-        var isReachable = true
-        switch reachability.connection {
-        case .none:
-            isReachable = false
-        default:
-            break
-        }
-
-        for watchers in networkStatusWatchers {
-            watchers.onNetworkAvailabilityStatusChanged(isEndpointReachable: isReachable)
-        }
-    }
-
-    var isNetworkReachable: Bool {
-        guard let reachability = reachability else {
-            return false
-        }
-
-        switch reachability.connection {
-        case .none:
-            return false
-        case .wifi:
-            return true
-        case .cellular:
-            return allowsCellularAccess
-        }
-    }
-
-    func isEligibleForExecution(_ operation: AWSAppSyncGraphQLOperation) -> Bool {
-        switch operation {
-        case .mutation:
-            return isNetworkReachable
-        case .query:
-            return true
-        case .subscription:
-            return true
-        }
-    }
-}
-
 public enum AWSAppSyncClientError: Error, LocalizedError {
     case requestFailed(Data?, HTTPURLResponse?, Error?)
     case noData(HTTPURLResponse)
